Move lib.rs to main.rs
